backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
black==23.11.0
flake8==6.1.0
fido2==1.1.2
pytz==2023.3
slowapi==0.1.9
maxminddb==2.6.2
# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0
pytest-mock==3.12.0
